Remove mapgen.h include from headers
