+++ src/mapgen.h
@@ -22,6 +22,7 @@
 #include "flexbuffer_json.h"
 #include "jmapgen_flags.h"
 #include "mapgen_parameter.h"
+#include "mapgen_primitives.h"
 #include "memory_fast.h"
 #include "point.h"
 #include "ret_val.h"
@@ -84,32 +85,6 @@ class mapgen_function_builtin : public virtual mapgen_function
 
 /////////////////////////////////////////////////////////////////////////////////
 ///// json mapgen (and friends)
-/*
- * Actually a pair of integers that can rng, for numbers that will never exceed INT_MAX
- */
-struct jmapgen_int {
-    int16_t val;
-    int16_t valmax;
-    explicit jmapgen_int( int v ) : val( v ), valmax( v ) {
-        cata_assert( v <= std::numeric_limits<int16_t>::max() );
-    }
-    jmapgen_int( int v, int v2 ) : val( v ), valmax( v2 ) {
-        cata_assert( v <= std::numeric_limits<int16_t>::max() );
-        cata_assert( v2 <= std::numeric_limits<int16_t>::max() );
-    }
-    /**
-     * Throws as usually if the json is invalid or missing.
-     */
-    jmapgen_int( const JsonObject &jo, std::string_view tag );
-    /**
-     * Throws if the json is malformed (e.g. a string not an integer, but does not throw
-     * if the member is just missing (the default values are used instead).
-     */
-    jmapgen_int( const JsonObject &jo, std::string_view tag, const int &def_val,
-                 const int &def_valmax );
-
-    int get() const;
-};
 
 /** Mapgen pieces will be applied in order of phases.  The phases are as
  * follows: */
@@ -201,11 +176,6 @@ struct jmapgen_setmap {
      * on the same map. Returns the name of the vehicle if a collision, and an empty string otherwise.
      **/
     ret_val<void> has_vehicle_collision( const mapgendata &dat, const tripoint_rel_ms &offset ) const;
-};
-
-struct spawn_data {
-    std::map<itype_id, jmapgen_int> ammo;
-    std::vector<point_rel_ms> patrol_points_rel_ms;
 };
 
 /**
